requirements.txt
# Core ML/NLP Libraries
transformers>=4.44.0
torch>=2.0.0
nltk>=3.8.1

# Gradio UI
gradio>=4.36.0

# Hugging Face
huggingface_hub>=0.25.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Visualization
pyvis>=0.3.2

# NER & Text Processing (minimal - no heavy model)
spacy>=3.7.0

# Google Gemini API
google-generativeai>=0.3.0

# ML Utils
scikit-learn>=1.3.0

# Environment
python-dotenv>=1.0.0

# Optional: Only needed if processing new data from scratch
# peft>=0.8.0
# trl>=0.9.0
# bitsandbytes>=0.43.0
# evaluate>=0.4.2
# datasets>=2.0.0

# Optional: Only if generating new character networks
# https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl